Commitment

+++ FinalCountdown/src/db/Users.sql
@@ -6,9 +6,9 @@ CREATE TABLE `users`(
 `id` int(11) NOT NULL AUTO_INCREMENT, 
 `first_name` varchar(255),
 `last_name` varchar(255),
-`initials` varchar (255),
+`initials` varchar(255),
 `password` varchar(255),
-`balance` int (10),
+`balance` double (10),
 PRIMARY KEY (`id`)) AUTO_INCREMENT = 4;
 
 INSERT INTO `users` VALUES ('1','Mathias', 'Højgaard', 'maho14am', '1234', 1.0), 
@@ -20,9 +20,9 @@ DROP TABLE IF EXISTS `admin`;
 CREATE TABLE `admin`(
 `id` int(11) NOT NULL AUTO_INCREMENT, 
 `name` varchar(255),
-`initials` varchar (255),
+`initials` varchar(255),
 `password` varchar(255),
-`currency` int(10),
+`currency` double(10),
 PRIMARY KEY (`id`)) AUTO_INCREMENT = 1;
 
 INSERT INTO `admin` VALUES ('1', 'CBS_ADMIN', 'Adm', '1234', 150);
